[SHIBUI-701]
Added support for resolver initialization. Each resolver now contains a
doInitialization boolean that controls whether the resolver's init
method should be run and if it should be added to the chaining resolver.
Currently, this only affects the filesystem metadata resolver's init method.
However, it affects all resolvers with respect to the chaining resolver.
Also added supporting tooltip and label and an update to the filesystem
resolver's json schema.

By default, doInitialization is true for all resolvers but the filesystem
resolver.