[SHIBUI-1029]
Updated backend app props to point spring mail at "mailhog" for use in
the pac4j docker build. Added dev config to pac4j docker-compose so
there will be admins to send email to. Added mailhog to the pac4j
docker-compose so that there is a place for the app to send new user
emails to.