[SHIBUI-1058][SHIBUI-1231]
Fixed a bug where we were persisting boolean relying party overrides
improperly. We now only persist when the value from the UI is true.
Note that we still persist based on how the override is defined in the
yaml. In other words, the UI controls whether or not to persist and the
yaml controls what is actually persisted.