1743
Added two new tests for 1743 and slowed down a test for 1742 a bit to
make it more stable. Note that 1743-1 has a lot of Selenese+IDE
weirdness where the "blank" field will somehow repopulate itself. This
issue is not reproducable manually. The test works around this by
setting the field to contain only one space.